requirements.txt
# conda
torch==1.12.1
Shapely==1.7.0
# pip
numpy==1.23.4
torchvision
pandas
addict
scikit-learn
opencv-python
tensorboard
pytorch_warmup
scikit-image
tqdm
p_tqdm
timm==0.4.12
imgaug>=0.4.0
ujson==1.35
yapf
albumentations==0.4.6
mmcv-full == 1.7.0
pathspec
# conda install pytorch==1.12.1 torchvision==0.13.1 torchaudio==0.12.1 cudatoolkit=11.3 -c pytorch